retos/web/Navaja_Wallet/deploy/app/requirements.txt
Flask==3.1.1
PyJWT==2.10.1
requests==2.32.4
gunicorn==23.0.0
